Rebis OK     Converted Parser's arrays of pointers to members handling $/@     syntax to unordered_maps;     The lexical scanner no longer supports the _$ syntax.     Type conversions to called functions are automatically performed     but from rhs assignments not: a typecast is required because     the compiler cannot determine the user's intentions. See     parser/inc/typename for an example.     Type conversions are also strict: when expecting an int, but     size_t is a polymorphic semantic value type a typecast is also     necessary.

+++ bisonc++/parser/setprecedence.cc
@@ -1,6 +1,6 @@
 #include "parser.ih"
 
-void Parser::setPrecedence(int type)
+void Parser::setPrecedence(size_t type)
 {
     Symbol *sp = 0;     // to prevent `sp uninitialized' warning by the
                         // compiler
